+++ options/index.html
@@ -41,6 +41,7 @@
                 </div>
             </div>
         </nav>
+        <div id="joorney-manifest" class="text-mute"></div>
 
         <div class="flex-fill d-none">
             <div id="joorney-announcement" class="alert alert-info rounded mx-5 mt-3">
