+++ options/pages/version/index.js
@@ -1,7 +1,8 @@
-import { SUPPORTED_VERSION, baseSettings } from '../../../configuration.js';
+import { baseSettings } from '../../../configuration.js';
 import { stringToHTML } from '../../../src/html_generator.js';
 import { includeVersion } from '../../../src/utils/authorize.js';
 import { StorageSync } from '../../../src/utils/browser.js';
+import { SUPPORTED_VERSION } from '../../../src/utils/version.js';
 
 export async function loadPage(features, currentSettings) {
     loadSupportedOdoo(currentSettings, features);
